requirements.txt
Django==2.2
pkg-resources==0.0.0
psycopg2==2.8.5
python-dotenv==0.14.0
pytz==2020.1
SQLAlchemy==1.3.19
sqlparse==0.3.1
